Got everything working

+++ import_timeseries.py
@@ -19,41 +19,49 @@ class QGisLizardImporter(object):
 
     def __init__(self, username, password):
         self.groundwater = GroundwaterTimeSeriesAndLocations()
-        self.groundwater.username = username
-        self.groundwater.password = password
+        self.groundwater.locs.username = username
+        self.groundwater.locs.password = password
+        self.groundwater.ts.username = username
+        self.groundwater.ts.password = password
 
     def download(self, south_west, north_east, start, end, groundwater_type):
         self.groundwater.bbox(south_west, north_east, start, end,
                                 groundwater_type)
         self.data = self.groundwater.results_to_dict()
-        pprint(self.data)
+        # pprint(self.data)
 
     def data_to_shape(self, directory="", filename="ggmn_groundwater.shp",
                       overwrite=False):
         # set up the shapefile driver
         driver = ogr.GetDriverByName("ESRI Shapefile")
+        print("Driver: %s" % driver)
 
         # create the data source
         self.file_location = os.path.join(directory, filename)
+        print("File location: %r" % self.file_location)
+
+        if os.path.exists(self.file_location):
+            if overwrite:
+                os.remove(self.file_location)
+            else:
+                raise WriteShapefileError(
+                    "File %s already exists, remove/rename it first" % self.file_location)
+
         if not os.path.exists(directory):
             os.makedirs(directory)
         data_source = driver.CreateDataSource(self.file_location)
+        print("data_source: %r" % data_source)
+        if data_source is None:
+            raise WriteShapefileError(
+                "Could not create shapefile %s" % self.file_location)
 
         # create the spatial reference, WGS84
         spatial_reference = osr.SpatialReference()
         spatial_reference.ImportFromEPSG(4326)
 
         # create the layer
-        try:
-            layer = data_source.CreateLayer("ggmn_groundwater",
-                                            spatial_reference, ogr.wkbPoint)
-        except AttributeError:
-            if overwrite:
-                os.remove(self.file_location)
-                layer = data_source.CreateLayer("ggmn_groundwater",
-                                                spatial_reference, ogr.wkbPoint)
-            else:
-                raise WriteShapefileError('Shapefile allready exsists!')
+        layer = data_source.CreateLayer("ggmn_groundwater",
+                                        spatial_reference, ogr.wkbPoint)
 
         # Add the fields
         field_name = ogr.FieldDefn("name", ogr.OFTString)
@@ -77,15 +85,15 @@ class QGisLizardImporter(object):
             # create the feature
             feature = ogr.Feature(layer.GetLayerDefn())
             # Set the attributes using the values from the delimited text file
-            feature.SetField("name", row['name'])
-            feature.SetField("loc_UUID", row['timeseries uuid'])
-            feature.SetField("ts_UUID", uuid)
-            feature.SetField("longitude", row['coordinates'][0])
-            feature.SetField("latitude", row['coordinates'][1])
-            feature.SetField("min", row['min'])
-            feature.SetField("mean", row['mean'])
-            feature.SetField("max", row['max'])
-            feature.SetField("range", row['range'])
+            feature.SetField("name", str(row['name']))
+            feature.SetField("loc_UUID", str(row['timeseries uuid']))
+            feature.SetField("ts_UUID", str(uuid))
+            feature.SetField("longitude", str(row['coordinates'][0]))
+            feature.SetField("latitude", str(row['coordinates'][1]))
+            feature.SetField("min", str(row['min']))
+            feature.SetField("mean", str(row['mean']))
+            feature.SetField("max", str(row['max']))
+            feature.SetField("range", str(row['range']))
 
 
             # create the WKT for the feature using Python string formatting
